re-styled what I had added before the imageError (footer, hero and positioning of the post content.
