+++ package.json
@@ -28,7 +28,8 @@
     "gatsby-transformer-remark": "^2.12.0",
     "gatsby-transformer-sharp": "^2.8.0",
     "react": "^16.12.0",
-    "react-dom": "^16.12.0"
+    "react-dom": "^16.12.0",
+    "react-scroll": "^1.8.1"
   },
   "devDependencies": {
     "env-cmd": "^10.1.0",
