+++ src/components/footer.module.css
@@ -4,42 +4,58 @@
     height:100vh; 
     display:grid;
     grid-template-columns: repeat(12, 1fr);
+    scroll-snap-align: start;
 }
-.footer h1{
+.footer h1, .footer a{
     color:white;
 }
+
+.footer a:hover{
+    color:#DD724B;
+}
+
+
+
+.footer h2{
+    font-size: 22px;
+    letter-spacing: 0px;
+}
+.footer ul{
+    list-style: none;
+}
+
+.footer p, .footer li{
+    font-size: 18px;
+}
+
 .logo{
-    grid-column: 4 / 9;
+    grid-row: 1;
+    grid-column: 5 / 9;
+    margin:auto;
 }
 
 .about{
+    grid-row: 2;
     grid-column: 2 / 6;
 }
-
-.team{
-    grid-column: 2 / 12;
+.tort {
+    grid-row: 2;
+    grid-column: 8 / 10;
 }
 
 .column1{
-    grid-column: 2 / 5;
+    grid-row: 3;
+    grid-column: 2 / 4;
 }
 
 .column2{
-    grid-column: 6 / 9;
+    grid-row: 3;
+    grid-column: 5 / 8;
 }
 
 
 .column3{
-    grid-column: 9 / 13;
-}
-
-
-
-.team h2{
-    font-size: 22px;
-    letter-spacing: 0px;
-}
-.team ul{
-    list-style: none;
+    grid-row: 3;
+    grid-column: 8 / 10;
 }
