+++ src/components/header.module.css
@@ -2,7 +2,6 @@
     width:100vw;
     display: flex;
     justify-content: center;
-    height:10vh;
 }
 
 h1{
@@ -10,7 +9,7 @@ h1{
     text-align: center;
     margin-top:auto;
     margin-bottom:auto;
-    font-style: 72px;
+    font-size: 4rem;
     letter-spacing: -1px;
 }
