+++ src/components/hero.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import {Link, graphql, useStaticQuery} from "gatsby"
-import spread from "./../images/spread.jpg" 
+import spread from "./../images/spread.jpg";
+import Header from '../components/header.js';
 import heroStyles from './hero.module.css';
 
 
@@ -8,7 +9,9 @@ import heroStyles from './hero.module.css';
 const Hero = () => {
     return(
         <div className={heroStyles.hero}>
-        <img src={spread} className={heroStyles.spread} alt="example spread" />
+            <Header />
+            <img src={spread} className={heroStyles.spread} alt="example spread" width="700px"/>
+            <a className={heroStyles.purchase} href="/"><h2>Purchase</h2></a>
         </div>
     );
 }
