--- src/components/posts/button.js
@@ -1,13 +1,0 @@
-
-// import React from 'react';
-// import buttonStyles from './button.module.css';
-
-// const Button = ({toggler}) => {
-//       return (
-//         <button className={buttonStyles.Button}
-//           onClick={toggler}><h2 className="readMore">Read more</h2></button>
-//       );
-//     }
-  
-   
-//   export default Button;
