--- src/components/posts/button.module.css
@@ -1,7 +1,0 @@
-Button{
-    border: solid 1px black;
-    padding:20px 50px;
-    display:inline-block;
-    font-family: Spectral;
-    background-color: white;
-}
