--- src/components/posts/posts.css
@@ -1,83 +1,0 @@
-.posts {
-    display:grid;
-    grid-template-columns: repeat(2, 1fr);
-}
-
-.post{
-    display:flex;
-    grid-column: 1 /3;
-    text-align: center;
-    height: 100vh;
-    position: relative;
-    scroll-snap-align: start;
-}
-.postDescrip, .postImg{
-    width:50vw;
-}
-
-.postDescrip{
-    justify-content: space-around;
-    display:flex;
-    flex-direction: column;
-    justify-content: space-around;
-}
-
-.postCopy{
-    background-color:#506944;
-    position: absolute;
-    top: 0;
-    right: 0;
-    height:100vh;
-    width: 50vw;
-    z-index: 9999;
-    padding:3rem 5rem;
-    /* transition: transform 2s ease;  */
-    animation: fadeIn ease 0.5s;
-}
-.postCopy h3{
-    font-weight: 400;
-    max-width: 25vw;
-    margin: 0 auto 1.5rem auto;
-}
-
-.copy{
-    text-align: left;
-}
-  .hide {
-    background-color:#506944;
-    position: absolute;
-    top: 0;
-    right: 0;
-    height:100vh;
-    width: 50vw;
-    z-index: 9999;
-    padding:4rem;
-    opacity:0;
- }
-
-
- @keyframes fadeIn {
-    0% {
-      opacity:0;
-    }
-    100% {
-      opacity:1;
-    }
-  }
-  
-
-button{
-    margin:0 auto;
-    border: solid 1px black;
-    width: 30vw;
-    padding:20px;
-    display:inline-block;
-    font-family: Spectral;
-    background-color: white;
-    outline: none;
-}
-
-button:hover {
-    background-color: #F9F9F9;
-}
-
