+++ src/pages/main.css
@@ -3,8 +3,11 @@
     padding:0;
     box-sizing: border-box;
 }
-
 @import url('https://fonts.googleapis.com/css2?family=Spectral:ital,wght@0,200;0,300;0,400;1,200;1,300;1,400&display=swap');
+
+html {
+    scroll-snap-type: y mandatory;
+}
 
 body{
     font-family: 'Spectral', serif;
@@ -21,7 +24,6 @@ h1, h2, h3{
 }
 
 
-
 Link, a {
     text-decoration: none;
 }
@@ -32,17 +34,17 @@ Link:visited, a:visited{
 
 
 p{
-    font-size: 20px;
+    font-size: 1rem;
     letter-spacing: -1px;
 }
 
 h2{
-    font-size: 72px;
-    letter-spacing: -5px;
+    font-size: 3.5rem;
+    letter-spacing: -3px;
+    max-width: 35vw;
 }
 
 h3{
     font-family: Helvetica neue, sans-serif;
-    font-weight:200;
     font-size: 18px;
 }
